Now Mozart compile mostly warning-free again! + Fixed most warnings about missing const with literal strings (remaining one is for virtualString to CString which is to be reviewed) + Changed the reporting flags in the DSS to sane values (removed -Weffc++ and -Winline, forced the use of most flags for GCC > 3) + Fixed most of the remaining warnings but   - usage of tmpnam -> requires some user-visible changes   - warnings in ozbison; well it's in code that's more than 12(!) years old. The good way would be to update the underlying bison. Pain ahead. + Fixed some bugs here and there (warnings are supposed to help locate them when they aren't drown in noise) +The output of make bootstrap dropped from 1.7 Mb to 189 kb!

+++ platform/dss/include/dss_comService.hh
@@ -74,7 +74,7 @@ protected:
   virtual ~DssReadBuffer() {}
 
 public:
-  virtual int availableData() const = 0;
+  virtual size_t availableData() const = 0;
   virtual bool canRead(size_t wanted) const = 0;
 
   // Read out data which should be deserialized. Assumes available
@@ -90,7 +90,7 @@ protected:
   virtual ~DssWriteBuffer() {}
 
 public:
-  virtual int availableSpace() const = 0;
+  virtual size_t availableSpace() const = 0;
   virtual bool canWrite(size_t wanted) const = 0;
 
   // ZACHARIAS: Check available space before writing, I know it did
